Refactoring and enforcing name rules on Evidence
