+++ eap_backend/eap_api/view_utils.py
@@ -1,14 +1,15 @@
 import functools
-import warnings
 from typing import Any, Callable, Optional, Union, cast
 
-from django.db.models import Q
 from django.db.models.query import QuerySet
 from django.forms.models import model_to_dict
 from django.http import JsonResponse
 from rest_framework.serializers import ReturnDict
 
-from . import models
+from .model_utils import (
+    get_case_property_claims,
+    traverse_child_property_claims,
+)
 from .models import (
     AssuranceCase,
     CaseItem,
@@ -20,78 +21,10 @@ from .models import (
     TopLevelNormativeGoal,
 )
 from .serializers import (
-    AssuranceCaseSerializer,
-    ContextSerializer,
-    EvidenceSerializer,
-    PropertyClaimSerializer,
+    TYPE_DICT,
     SandboxSerializer,
-    StrategySerializer,
-    TopLevelNormativeGoalSerializer,
+    get_case_id,
 )
-
-TYPE_DICT = {
-    "assurance_case": {
-        "serializer": AssuranceCaseSerializer,
-        "model": AssuranceCase,
-        "children": ["goals"],
-        "fields": (
-            "name",
-            "description",
-            "lock_uuid",
-            "owner",
-            "color_profile",
-        ),
-    },
-    "goal": {
-        "serializer": TopLevelNormativeGoalSerializer,
-        "model": TopLevelNormativeGoal,
-        "children": ["context", "property_claims", "strategies"],
-        "fields": (
-            "name",
-            "short_description",
-            "long_description",
-            "keywords",
-        ),
-        "parent_types": [("assurance_case", False)],
-    },
-    "context": {
-        "serializer": ContextSerializer,
-        "model": Context,
-        "children": [],
-        "fields": ("name", "short_description", "long_description"),
-        "parent_types": [("goal", False)],
-    },
-    "property_claim": {
-        "serializer": PropertyClaimSerializer,
-        "model": PropertyClaim,
-        "children": ["property_claims", "evidence"],
-        "fields": ("name", "short_description", "long_description"),
-        "parent_types": [
-            ("goal", False),
-            ("property_claim", False),
-            ("strategy", False),
-        ],
-    },
-    "strategy": {
-        "serializer": StrategySerializer,
-        "model": Strategy,
-        "children": ["property_claims"],
-        "fields": ("name", "short_description", "long_description"),
-        "parent_types": [
-            ("goal", False),
-        ],
-    },
-    "evidence": {
-        "serializer": EvidenceSerializer,
-        "model": Evidence,
-        "children": [],
-        "fields": ("name", "short_description", "long_description", "URL"),
-        "parent_types": [("property_claim", True)],
-    },
-}
-# Pluralising the name of the type should be irrelevant.
-for k, v in tuple(TYPE_DICT.items()):
-    TYPE_DICT[k + "s" if not k.endswith("y") else k[:-1] + "ies"] = v
 
 
 class SandboxUtils:
@@ -354,9 +287,7 @@ class UpdateIdentifierUtils:
             (
                 top_level_claim_ids,
                 child_claim_ids,
-            ) = UpdateIdentifierUtils._get_case_property_claims(
-                current_case_goal, current_case_strategies
-            )
+            ) = get_case_property_claims(current_case_goal, current_case_strategies)
 
             UpdateIdentifierUtils._update_sequential_identifiers(
                 Evidence.objects.filter(
@@ -377,7 +308,7 @@ class UpdateIdentifierUtils:
             )
 
             for _, property_claim in enumerate(parent_property_claims):
-                UpdateIdentifierUtils._traverse_child_property_claims(
+                traverse_child_property_claims(
                     lambda index, child, parent: UpdateIdentifierUtils._update_item_name(
                         child, f"{parent.name}.", index + 1
                     ),
@@ -386,34 +317,6 @@ class UpdateIdentifierUtils:
 
             if model_instance is not None:
                 model_instance.refresh_from_db()
-
-    @staticmethod
-    def _traverse_child_property_claims(
-        on_child_claim: Callable[[int, PropertyClaim, PropertyClaim], None],
-        parent_claim_id: int,
-    ):
-        """Applies a function to all the children of a Property Claim.
-
-        Args:
-            on_child_claim: The function to call on each child claim.
-            parent_claim_id: The id of the claim we will traverse.
-        """
-        child_property_claims = PropertyClaim.objects.filter(
-            property_claim_id=parent_claim_id
-        ).order_by("id")
-
-        if len(child_property_claims) == 0:
-            return
-        else:
-            for index, child_property_claim in enumerate(child_property_claims):
-                on_child_claim(
-                    index,
-                    child_property_claim,
-                    PropertyClaim.objects.get(pk=parent_claim_id),
-                )
-                UpdateIdentifierUtils._traverse_child_property_claims(
-                    on_child_claim, child_property_claim.pk
-                )
 
     @staticmethod
     def _compare_property_claims(
@@ -438,38 +341,6 @@ class UpdateIdentifierUtils:
         return result
 
     @staticmethod
-    def _get_case_property_claims(
-        goal: TopLevelNormativeGoal, strategies: QuerySet
-    ) -> tuple:
-        """Retrieves all the property claims associated to a goal and a list of strategies.
-
-        Args:
-            goal: Goal whose property claims we will extract.
-            strategies: Strategies whose property claims we will extract.
-
-        Returns:
-            A tuple containing parent property claims and child property claims, all sorted
-            by primary key.
-        """
-        strategy_ids: list[int] = [strategy.pk for strategy in strategies]
-        top_level_claims: QuerySet = PropertyClaim.objects.filter(
-            Q(goal_id=goal.pk) | Q(strategy__id__in=strategy_ids)
-        )
-
-        top_level_claim_ids: list[int] = [claim.pk for claim in top_level_claims]
-
-        child_claim_ids: list[int] = []
-        for parent_claim_id in top_level_claim_ids:
-            UpdateIdentifierUtils._traverse_child_property_claims(
-                lambda _, child, parent: child_claim_ids.append(  # noqa: ARG005
-                    child.pk
-                ),
-                parent_claim_id,
-            )
-
-        return top_level_claim_ids, sorted(child_claim_ids)
-
-    @staticmethod
     def _update_item_name(case_item: CaseItem, prefix: str, number: int) -> None:
         """Updates the name of a case item, given a prefix and a sequence number."""
         case_item.name = f"{prefix}{number}"
@@ -480,29 +351,6 @@ class UpdateIdentifierUtils:
         """For a list of case items, it updates their name according to its order."""
         for model_index, model in enumerate(query_set):
             UpdateIdentifierUtils._update_item_name(model, prefix, model_index + 1)
-
-
-def get_case_id(item) -> Optional[int]:
-    """Return the id of the case in which this item is. Works for all item types."""
-    # In some cases, when there's a ManyToManyField, instead of the parent item, we get
-    # an iterable that can potentially list all the parents. In that case, just pick the
-    # first.
-    if hasattr(item, "first"):
-        item = item.first()
-
-    if isinstance(item, models.AssuranceCase):
-        return item.id
-    for _k, v in TYPE_DICT.items():
-        if isinstance(item, v["model"]):
-            for parent_type, _ in v["parent_types"]:
-                parent = getattr(item, parent_type)
-                if parent is not None:
-                    return get_case_id(parent)
-    # TODO This should probably be an error raise rather than a warning, but currently
-    # there are dead items in the database without parents which hit this branch.
-    msg = f"Can't figure out the case ID of {item}."
-    warnings.warn(msg)
-    return None
 
 
 def filter_by_case_id(items, request):
